refactor: 기존 코드 -> 새로 정한 패키지 구조에 맞추기
